+++ datafiles/formats.py
@@ -1,5 +1,6 @@
 import json
 from abc import ABCMeta, abstractmethod
+from io import StringIO
 from pathlib import Path
 from typing import IO, Any, Dict, List
 
@@ -71,7 +72,11 @@ class YAML(Formatter):
 
     @classmethod
     def serialize(cls, data):
-        text = yaml.round_trip_dump(data)
+        f = StringIO()
+        y = yaml.YAML()
+        y.indent(mapping=2, sequence=4, offset=2)
+        y.dump(data, f)
+        text = f.getvalue().strip() + '\n'
         return "" if text == "{}\n" else text