+++ tests/conftest.py
@@ -1,8 +1,7 @@
 import os
 from pathlib import Path
-from shutil import get_terminal_size, rmtree
+from shutil import rmtree
 
-import log
 import pytest
 
 from datafiles import settings
@@ -32,50 +31,3 @@ def create_tmp():
     if path.exists():
         rmtree(path)
     path.mkdir(exist_ok=True)
-
-
-@pytest.fixture
-def logbreak():
-    def _(message=""):
-        width = get_terminal_size().columns - 31
-        if message:
-            line = '-' * (width - len(message) - 1) + ' ' + message
-        else:
-            line = '-' * width
-        log.info(line)
-
-    return _
-
-
-@pytest.fixture(scope='session')
-def dedent():
-    return lambda text: text.replace(' ' * 4, '').strip() + '\n'
-
-
-# TODO: Move these utilities into the library
-
-
-@pytest.fixture(scope='session')
-def write(dedent):
-    def _(path: str, text: str) -> None:
-        _path = Path(path).resolve()
-        _text = dedent(text)
-        message = f'Writing file: {_path}'
-        log.info(message)
-        log.debug('=' * len(message) + '\n\n' + (_text or '<nothing>\n'))
-        _path.write_text(_text)
-
-    return _
-
-
-@pytest.fixture(scope='session')
-def read():
-    def _(path: str) -> str:
-        _path = Path(path).resolve()
-        message = f'Reading file: {_path}'
-        log.info(message)
-        text = _path.read_text()
-        log.debug('=' * len(message) + '\n\n' + (text or '<nothing>\n'))
-        return text
-
-    return _