+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed ORM methods for datafiles with relative path patterns.
 - Added plugin for `mypy` support.
+- Updated YAML format to indent lists.
 
 # 0.3 (2019-06-09)